+++ integration-test/exposure-test/src/main/java/com/truward/brikar/test/exposure/ExposureServerLauncher.java
@@ -9,7 +9,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.Callable;
 
 
 /**